+++ pyelastictest/tests/test_client.py
@@ -5,7 +5,7 @@ class TestExtendedClient(IsolatedTestCase):
 
     def _make_one(self):
         from pyelastictest import client
-        return client.ExtendedClient(self.es_cluster[0].address)
+        return client.ExtendedClient(self.es_cluster.urls)
 
     def test_create_template(self):
         client = self._make_one()